backend/requirements.txt

flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
stripe==7.11.0
